Conflicts:
	scilab/Makefile.in
	scilab/configure
	scilab/configure.ac
	scilab/contrib/Makefile.in
	scilab/libs/MALLOC/Makefile.in
	scilab/libs/Makefile.in
	scilab/libs/doublylinkedlist/Makefile.in
	scilab/libs/dynamiclibrary/Makefile.in
	scilab/libs/hashtable/Makefile.in
	scilab/libs/libst/Makefile.in
	scilab/modules/Makefile.in
	scilab/modules/action_binding/Makefile.in
	scilab/modules/api_scilab/Makefile.in
	scilab/modules/arnoldi/Makefile.in
	scilab/modules/atoms/Makefile.in
	scilab/modules/boolean/Makefile.in
	scilab/modules/cacsd/Makefile.in
	scilab/modules/call_scilab/Makefile.in
	scilab/modules/commons/Makefile.in
	scilab/modules/compatibility_functions/Makefile.in
	scilab/modules/completion/Makefile.in
	scilab/modules/console/Makefile.in
	scilab/modules/core/Makefile.in
	scilab/modules/core/includes/version.h
	scilab/modules/data_structures/Makefile.in
	scilab/modules/demo_tools/Makefile.in
	scilab/modules/development_tools/Makefile.in
	scilab/modules/development_tools/src/fake/Makefile.in
	scilab/modules/differential_equations/Makefile.in
	scilab/modules/double/Makefile.in
	scilab/modules/dynamic_link/Makefile.in
	scilab/modules/elementary_functions/Makefile.in
	scilab/modules/fftw/Makefile.in
	scilab/modules/fileio/Makefile.in
	scilab/modules/functions/Makefile.in
	scilab/modules/genetic_algorithms/Makefile.in
	scilab/modules/graph/Makefile.in
	scilab/modules/graphic_export/Makefile.in
	scilab/modules/graphics/Makefile.in
	scilab/modules/graphics/src/c/BuildObjects.c
	scilab/modules/graphics/src/c/getHandleProperty/get_user_data_property.c
	scilab/modules/gui/Makefile.in
	scilab/modules/hdf5/Makefile.in
	scilab/modules/helptools/Makefile.in
	scilab/modules/history_browser/Makefile.in
	scilab/modules/history_manager/Makefile.in
	scilab/modules/integer/Makefile.in
	scilab/modules/interpolation/Makefile.in
	scilab/modules/intersci/Makefile.in
	scilab/modules/io/Makefile.in
	scilab/modules/javasci/Makefile.in
	scilab/modules/jvm/Makefile.in
	scilab/modules/linear_algebra/Makefile.in
	scilab/modules/localization/Makefile.in
	scilab/modules/m2sci/Makefile.in
	scilab/modules/maple2scilab/Makefile.in
	scilab/modules/matio/Makefile.in
	scilab/modules/mexlib/Makefile.in
	scilab/modules/modules_manager/Makefile.in
	scilab/modules/optimization/Makefile.in
	scilab/modules/output_stream/Makefile.in
	scilab/modules/overloading/Makefile.in
	scilab/modules/parallel/Makefile.in
	scilab/modules/parameters/Makefile.in
	scilab/modules/polynomials/Makefile.in
	scilab/modules/pvm/Makefile.in
	scilab/modules/randlib/Makefile.in
	scilab/modules/renderer/Makefile.in
	scilab/modules/scicos/Makefile.in
	scilab/modules/scicos_blocks/Makefile.in
	scilab/modules/scinotes/Makefile.in
	scilab/modules/shell/Makefile.in
	scilab/modules/signal_processing/Makefile.in
	scilab/modules/simulated_annealing/Makefile.in
	scilab/modules/sound/Makefile.in
	scilab/modules/sparse/Makefile.in
	scilab/modules/special_functions/Makefile.in
	scilab/modules/spreadsheet/Makefile.in
	scilab/modules/statistics/Makefile.in
	scilab/modules/string/Makefile.in
	scilab/modules/symbolic/Makefile.in
	scilab/modules/tclsci/Makefile.in
	scilab/modules/texmacs/Makefile.in
	scilab/modules/time/Makefile.in
	scilab/modules/types/Makefile.in
	scilab/modules/ui_data/Makefile.in
	scilab/modules/umfpack/Makefile.in
	scilab/modules/windows_tools/Makefile.in
	scilab/modules/xcos/Makefile.in

+++ scilab/modules/completion/src/c/getfieldsdictionary.c
@@ -40,7 +40,7 @@ char **getfieldsdictionary(char *lineBeforeCaret, char *pattern, int *size)
     char **pstData = NULL;
     char *var = NULL;
     char *lineBeforePoint = NULL;
-    int pos = strlen(lineBeforeCaret) - strlen(pattern) - 1;
+    int pos = (int)(strlen(lineBeforeCaret) - strlen(pattern) - 1);
 
     if (pos <= 0 || lineBeforeCaret[pos] != '.')
     {
